pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with industry affliations"
authors = [
    {name = "avinash-pandeyanf",email = "avinash.pandey8900@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "typer (==0.9.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.scripts]
get-papers-list = "pubmed_fetcher.cli:app"